Updated amount validation of CreatePaymentDialog
